Move UpdateInterested to FirebaseUtils
